refactor : 야구 게임 레이어의 이름 변경
게임을 진행하는 레이어와 게임이 종료되었을때의 레이어를 나누었기에 원래의 레이어 이름 변경
