+++ src/test/java/baseball/ApplicationTest.java
@@ -1,7 +1,7 @@
 package baseball;
 
-import baseball.Game.BaseBallGame;
-import baseball.Game.BaseBallGameImpl;
+import baseball.ProcessGame.BaseBallGame;
+import baseball.ProcessGame.BaseBallGameImpl;
 import camp.nextstep.edu.missionutils.test.NsTest;
 import org.junit.jupiter.api.Test;
